# openapi.yaml
openapi: 3.0.3
info:
  title: Whop SaaS Marketplace API
  description: |
    Production-grade marketplace API enabling multi-tenant SaaS, Stripe Connect payments,
    license management, and webhook event broadcasting.

    ## Authentication
    All endpoints require authentication via session token cookie (`session_token`).

    ## Multi-Tenancy
    All resources are scoped to companies. Users can access multiple companies with different roles.

    ## Rate Limiting
    - 100 requests per minute per IP
    - Returns `429 Too Many Requests` when exceeded

    ## Security
    - Tenant isolation via `company_id` filtering
    - Role-based access control (owner/admin/member)
    - Money stored in minor units (cents) to avoid floating-point errors
  version: 1.0.0
  contact:
    name: API Support
    email: api@whopsaas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.whopsaas.com
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: JWT session token (HTTP-only cookie)
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          example: whk_abc123
        company_id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - payment.succeeded
              - payment.failed
              - membership.created
              - membership.updated
              - membership.canceled
              - product.created
              - product.updated
              - product.deleted
        api_version:
          type: string
          default: v1
        secret:
          type: string
          description: HMAC secret (masked after creation)
          example: whsec_abc123...
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            deliveries:
              type: integer
    WebhookCreate:
      type: object
      properties:
        company_id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          minItems: 1
        api_version:
          type: string
          default: v1
        active:
          type: boolean
          default: true
      required:
        - company_id
        - url
        - events
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /api/webhooks:
    get:
      tags:
        - Webhooks
      summary: List webhooks
      description: List webhook subscriptions (admins only)
      parameters:
        - name: company_id
          in: query
          required: true
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Forbidden (admins only)

    post:
      tags:
        - Webhooks
      summary: Create webhook
      description: Create webhook subscription (admins only). Returns full secret only on creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Webhook'
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: Full HMAC secret (save securely, won't be shown again)
                      warning:
                        type: string
        '403':
          description: Forbidden
...